a couple pedantic changes and makes factory file separate. We can deal with the extra files locations when we tackle #87
